+++ demo/api/src/pages/dto/page.input.ts
@@ -1,20 +1,20 @@
 import { BlockInputInterface } from "@comet/blocks-api";
+import { RootBlockInputScalar } from "@comet/cms-api";
 import { Field, InputType } from "@nestjs/graphql";
 import { Transform } from "class-transformer";
 import { ValidateNested } from "class-validator";
-import { GraphQLJSONObject } from "graphql-type-json";
 
 import { PageContentBlock } from "../blocks/PageContentBlock";
 import { SeoBlock } from "../blocks/seo.block";
 
 @InputType()
 export class PageInput {
-    @Field(() => GraphQLJSONObject)
+    @Field(() => RootBlockInputScalar(PageContentBlock))
     @Transform(({ value }) => PageContentBlock.blockInputFactory(value), { toClassOnly: true })
     @ValidateNested()
     content: BlockInputInterface;
 
-    @Field(() => GraphQLJSONObject)
+    @Field(() => RootBlockInputScalar(SeoBlock))
     @Transform(({ value }) => SeoBlock.blockInputFactory(value), { toClassOnly: true })
     @ValidateNested()
     seo: BlockInputInterface;